Fix double free in proc_pid module
The original code called `.as_ptr()` (which *borrows* its argument) and
passed it to `Vec::from_raw_parts()` (which *consumes* it). So the
underlying buffer would have ended up with two owners, which is wrong.

This patch changes the code to use `.set_len()` instead, which borrows
the allocation instead of consuming it. It also uses `.as_mut_ptr()`
instead of `.as_ptr()` for const-correctness.

Closes #2