Use bindgen for OSX bindings
Manually written library bindings turn out to be fragile in the event
that the the underlying library is updated. We have no control over
the user's systems, so we cannot assume that the bindings we write
will remain valid forever.

This change introduces bindgen to create rust bindings using
headers from the system on which the package is built.